pyproject.toml
[tool.poetry]
name = "upper-realworld"
version = "0.1.0"
description = ""
authors = ["Rafael Zanatta <rafael.zanatta@upperds.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
SQLAlchemy = "^1.4.23"
alembic = "^1.7.1"
uvicorn = "^0.15.0"
databases = "^0.5.2"
asyncio = "^3.4.3"
asyncpg = "^0.24.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
httpx = "^0.19.0"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
pytest-env = "^0.6.2"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
extend-exclude = "app/db/migrations"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true